Dockerfile Optimization (RUN concatenation)
Made some simple concatenation this should work but better doublecheck. This way we saved a couple of layers in the build, also why using amazonlinux:2018 instead of the latest? 
Also used ENTRYPOINT instead of CMD so this way you may run the  container without the need to specify anything so it's simple like:
docker run -it --rm mreferre/eksutils